backend/requirements.txt
faiss-cpu
groq
gunicorn
langchain-groq
PyPDF2
chromadb==0.5.5
chroma-migrate
schedule
langchain_google_genai
langchain
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
Flask
flask-cors